requirements/base.txt
# Base requirements for CodebeGen

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & Machine Learning
torch==2.1.1
transformers==4.36.0
accelerate==0.25.0
tokenizers==0.15.0
huggingface-hub==0.19.4
bitsandbytes==0.41.3

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
jinja2==3.1.2
rich==13.7.0

# Rate Limiting
slowapi==0.1.9

# Task Queue (Optional - for background processing)
celery==5.3.4
redis==5.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
